Add basic tests in anticipation of Ruby 1.9 fix
In order to verify the current behavior, I have added a super basic
test suite that pipes an included dummy Xcode project into
ocunit2junit.rb. This isn't the best system (it's slow, hitting the
filesystem, and brittle), but it'll do (for me) for now when making
OCUnit2JUnit compatible with Ruby 1.9.